+++ core/src/main/java/cucumber/table/xstream/DataTableWriter.java
@@ -1,8 +1,0 @@
-package cucumber.table.xstream;
-
-import com.thoughtworks.xstream.io.AbstractWriter;
-import cucumber.table.DataTable;
-
-public abstract class DataTableWriter extends AbstractWriter {
-    public abstract DataTable getDataTable();
-}